requirements.txt
snowflake-connector-python
cryptography
streamlit
python-dotenv